LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/selection-direction-on-single-click-expected.txt

hello, world

FAIL direction returns "none" when the selection is collapsed assert_equals: expected Text node "hello, world" but got null
